Add OutBoundQueueSpec to hspec-discovery
