+++ node-sketch.cabal
@@ -239,6 +239,7 @@ test-suite node-sketch-test
   main-is:             Test.hs
   other-modules:       Spec
                        Test.NodeSpec
+                       Test.Network.Broadcast.OutboundQueueSpec
                        Test.Util
   type:                exitcode-stdio-1.0
   build-depends:       base >= 4.8 && < 5
