+++ test/Test.hs
@@ -1,6 +1,5 @@
-import           Test.NodeSpec (spec)
+import           Spec       (spec)
 import           Test.Hspec (hspec)
-import           Test.Util (makeTCPTransport, makeInMemoryTransport)
 
 main :: IO ()
 main = hspec spec
